Remove reliance on import from derivation
`callCabal2nix` runs `cabal2nix` internally to obtain a package
description for a given Cabal file. Since this uses IFD, the single
threaded evaluator has to stop until that build has finished. This is
usually a slight annoyance, but can in the worst case require to build
GHC if binary cache is lacking.

We can avoid this by simply manually generating the file and checking it
in.