nix/sources.json: switch to nixos-unstable
Gives us up to date dependencies and a modern GHC in the devshell.
As #52 is not merged yet, we'll update to a recent channel revision
that still has GHC 9.0 and LTS-19, so the repository remains buildable.
We'll just use the default nodejs from now on, as likely no one is going
to update this pin in the future…